Analogue input!

+++ common/src/main/java/io/github/foundationgames/automobility/block/entity/AutomobileAssemblerBlockEntity.java
@@ -319,7 +319,7 @@ public class AutomobileAssemblerBlockEntity extends BlockEntity implements Rende
     }
 
     @Override
-    public int getTurboCharge() {
+    public float getTurboCharge() {
         return 0;
     }
